requirements.txt
streamlit==1.32.0
gspread==6.1.0
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
oauth2client==4.1.3
python-dotenv==1.0.1
reportlab==4.1.0
svglib==1.5.1
pandas==2.2.1
